requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-groq
langchain_core
fastapi
uvicorn
langserve
arxiv
wikipedia
ddgs








